# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MTCBackend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Appointment": {
      "get": {
        "tags": [
          "Appointment"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Appointment"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Appointment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AppointmentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Appointment"
            }
          }
        }
      }
    },
    "/api/Appointment/{id}": {
      "get": {
        "tags": [
          "Appointment"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Appointment"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Appointment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Appointment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Appointment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Billing": {
      "post": {
        "tags": [
          "Billing"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BillDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Bill"
            }
          }
        }
      }
    },
    "/api/Billing/{billId}/payment": {
      "post": {
        "tags": [
          "Billing"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "billId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PaymentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Billing/{id}": {
      "get": {
        "tags": [
          "Billing"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Bill"
            }
          }
        }
      }
    },
    "/api/Billing/ledger/{patientId}": {
      "get": {
        "tags": [
          "Billing"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "patientId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Ledger"
            }
          }
        }
      }
    },
    "/api/DigitalSignature/capture": {
      "post": {
        "tags": [
          "DigitalSignature"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "formId",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "patientId",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DigitalSignature/verify/{signatureId}": {
      "post": {
        "tags": [
          "DigitalSignature"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "signatureId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/File/upload": {
      "post": {
        "tags": [
          "File"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "formId",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "patientId",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/File/download/{fileId}": {
      "get": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Form/{formId}": {
      "get": {
        "tags": [
          "Form"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "formId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FormDTO"
            }
          }
        }
      }
    },
    "/api/Form/patient/{patientId}": {
      "get": {
        "tags": [
          "Form"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "patientId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FormDTO"
              }
            }
          }
        }
      }
    },
    "/api/Form": {
      "post": {
        "tags": [
          "Form"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FormDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Form/submission": {
      "post": {
        "tags": [
          "Form"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FormSubmissionDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Patient/{id}/dashboard": {
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userType"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Appointment": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "appointmentDate": {
          "format": "date-time",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "patientId": {
          "format": "uuid",
          "type": "string"
        },
        "patient": {
          "$ref": "#/definitions/User"
        },
        "staffId": {
          "format": "uuid",
          "type": "string"
        },
        "staff": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "AppointmentDTO": {
      "type": "object",
      "properties": {
        "appointmentDate": {
          "format": "date-time",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "patientId": {
          "format": "uuid",
          "type": "string"
        },
        "staffId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "Bill": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "patientId": {
          "format": "int32",
          "type": "integer"
        },
        "patient": {
          "$ref": "#/definitions/User"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "status": {
          "type": "string"
        },
        "lineItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LineItem"
          }
        },
        "payments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Payment"
          }
        }
      }
    },
    "BillDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "patientId": {
          "format": "int32",
          "type": "integer"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "status": {
          "type": "string"
        },
        "lineItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LineItemDTO"
          }
        }
      }
    },
    "FormDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "patientId": {
          "format": "uuid",
          "type": "string"
        },
        "fields": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FormFieldDTO"
          }
        }
      }
    },
    "FormFieldDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "fieldType": {
          "type": "string"
        },
        "fieldName": {
          "type": "string"
        }
      }
    },
    "FormSubmissionDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "formId": {
          "format": "uuid",
          "type": "string"
        },
        "patientId": {
          "format": "uuid",
          "type": "string"
        },
        "signatureId": {
          "format": "uuid",
          "type": "string"
        },
        "uploadedFiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UploadedFileDTO"
          }
        }
      }
    },
    "Ledger": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "patientId": {
          "format": "uuid",
          "type": "string"
        },
        "patient": {
          "$ref": "#/definitions/User"
        },
        "totalCharges": {
          "format": "double",
          "type": "number"
        },
        "totalPayments": {
          "format": "double",
          "type": "number"
        },
        "outstandingBalance": {
          "format": "double",
          "type": "number",
          "readOnly": true
        }
      }
    },
    "LineItem": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "billId": {
          "format": "int32",
          "type": "integer"
        },
        "bill": {
          "$ref": "#/definitions/Bill"
        },
        "description": {
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "LineItemDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "Payment": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "billId": {
          "format": "int32",
          "type": "integer"
        },
        "bill": {
          "$ref": "#/definitions/Bill"
        },
        "paymentDate": {
          "format": "date-time",
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "PaymentDTO": {
      "type": "object",
      "properties": {
        "amount": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "UploadedFileDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "filePath": {
          "type": "string"
        },
        "file": {
          "format": "binary",
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date-time",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/UserType"
        },
        "medicalRecordNumber": {
          "type": "string"
        },
        "staffPosition": {
          "type": "string"
        },
        "patientAppointments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Appointment"
          }
        },
        "staffAppointments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Appointment"
          }
        }
      }
    },
    "UserType": {
      "enum": [
        "Patient",
        "Staff"
      ],
      "type": "string"
    }
  }
}